Split type definitions into domain-specific files
