+++ src/types/api.ts
@@ -1,0 +1,9 @@
+/**
+ * API and service-related type definitions
+ */
+
+export interface ApiResponse<T> {
+  success: boolean;
+  data?: T;
+  error?: string;
+}
